Update kallistobustools_count.nf
instead of using only the first elements of the input array, it now joins the filenames in a single string usig whitespaces.

This would make it possible to use it when more replicates are available

We may want to add a checking so it complains when not using paired end data.