fix: STAR mtx conversion when using GeneFull
When STAR is run  with the flag `--soloFeatures GeneFull` (permits
counting of exonic and intronic reads), the output is stored in
`*.Solo.out/GeneFull/` and not `*.Solo.out/Gene`. As a result, matrix
conversion results in an error, as matrix, barcodes and features cannot
be found. This error can be fixed by adding an asterisk in the file
path provided to the mtx conversion modules.