Add GTF_GENE_FILTER to remove extraneous features
It is possible to use a GTF file that has annotations on non-chromosomal
sequences. If these non-chromosomal features are absent from the
genome_fasta provided to the pipeline, some downstream tools will break.

Here we remove any features that do not have a matching source
sequence in the genome_fasta.